feat: add downcast on `Sleep` trait
This commit adds `as_any` downcast method for the `Sleep` trait.

BREAKING CHANGE: `Sleep` trait now needs `as_any` implementation
